fix(matte): raw cond latents; fp16 default; preserve checkpoints on dev-sync
pipeline.py: _encode_latents must NOT divide by vae scaling_factor. The
VideoMaMa UNet is trained on raw VAE latents (matching okdalto's
VideoInferencePipeline, which multiplies by scaling_factor then divides it
back out -> net x1). Dividing over-scaled the conditioning ~5.5x and produced
a systematically blurry matte on every frame. Verified 5/5 vs the desktop /
vanilla reference on identical single-frame inputs.

nodes.py: harden the loader precision fallback so fp16 is the true default -
only an explicit "bf16" opts out; a stale "auto" widget now resolves to fp16
instead of bf16.

sync_to_dev.py: the rmtree+copytree mirror deleted matte/checkpoints/ (the
gitignored ~10GB SVD+VideoMaMa weights) on every matte sync, forcing a
re-download. Add PRESERVE_IN_DEST + selective _clean_dest so runtime-
downloaded model data survives a sync.

docs: add docs/maintainers/matte-pipeline.md (latent-scaling invariant,
black-output causes, fidelity levers, single-frame + vanilla-oracle debug
loop, checkpoint preservation).

+++ matte/nodes.py
@@ -97,7 +97,9 @@ class KoolookMatteLoader:
         enable_vae_tiling: bool,
         enable_vae_slicing: bool
     ):
-        weight_dtype = torch.float16 if precision == "fp16" else torch.bfloat16
+        # fp16 is the default; only an explicit bf16 opts out. Anything else
+        # (incl. a stale "auto" widget from an older node version) -> fp16.
+        weight_dtype = torch.bfloat16 if precision == "bf16" else torch.float16
         node_dir = Path(__file__).parent
         # relative paths resolve next to this pack; absolute paths pass through
         base = base_model_path if os.path.isabs(base_model_path) else str(node_dir / base_model_path)
